// frontend/src/types/api.generated.ts
export interface components {
  schemas: {
// ...
    UserRegistrationRequest: {
      /**
       * Email адрес
       * Format: email
       */
      email: string | null;
      password: string;
      password_confirm: string;
      /** Имя */
      first_name: string;
      /** Фамилия */
      last_name?: string;
      /** Номер телефона */
      phone?: string;
      role: components['schemas']['UserRegistrationRoleEnum'];
      /**
       * Название компании
       * @description Для B2B пользователей
       */
      company_name?: string;
      /**
       * ИНН
       * @description ИНН для B2B пользователей
       */
      tax_id?: string;
      /**
       * Страна
       * @description Страна регистрации B2B-клиента (для маршрутизации на специалиста)
       *
       *     * `Россия` - Россия
       *     * `Беларусь` - Беларусь
       *     * `Казахстан` - Казахстан
       */
      country?: components['schemas']['CountryEnum'];
      pdp_consent: boolean;
      /** @default false */
      marketing_consent: boolean;
      pdp_consent_text_version: string;
      /** @default  */
      marketing_consent_text_version: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}